Move _handle_config_value tests to test_unit_utils.py
...from test_unit_modify_fsurdat.py and add to them is_list = False test.
Also break up lonRange0To360 tests to a test per case.
